requirements.txt
Flask==3.0.0
Pillow==10.2.0
requests==2.31.0
ultralytics>=8.3.0
megadetector>=10.0.17
opencv-python-headless>=4.8.0
